Move `line_span` to source.rs
